スキーマの列属性の規則性を統一、およびマイグレーション用SQLを追加 fixes #187 /trunk/db/schema/mysql/migration/migration_r166.sql
git-svn-id: http://svn.l1j.org/l1j/trunk@166 698199e4-35c4-443b-82c4-7f4b4ae73c13

+++ db/schema/mysql/race_tickets.sql
@@ -4,10 +4,10 @@ SET FOREIGN_KEY_CHECKS=0;
 -- ----------------------------
 DROP TABLE IF EXISTS `race_tickets`;
 CREATE TABLE IF NOT EXISTS `race_tickets` (
-  `item_obj_id` int(11) NOT NULL,
-  `round` int(7) NOT NULL,
+  `item_obj_id` int(10) NOT NULL,
+  `round` int(10) NOT NULL,
   `allotment_percentage` double NOT NULL,
-  `victory` int(1) NOT NULL,
-  `runner_num` int(2) NOT NULL,
+  `victory` tinyint(1) NOT NULL,
+  `runner_num` tinyint(3) NOT NULL,
   PRIMARY KEY (`item_obj_id`,`round`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
